Make environment watched by separate controller
