+++ controllers/starter.go
@@ -128,6 +128,12 @@ func SetupAllReconcilers(mgr controllerruntime.Manager, cfg *config.OperatorConf
 		}).SetupWithManager(mgr); err != nil {
 			return err
 		}
+
+		if err = (&EnvironmentReconciler{
+			k8sClient: mgr.GetClient(),
+		}).SetupWithManager(mgr); err != nil {
+			return err
+		}
 	}
 
 	if cfg.EnableTokenUpload {
