+++ integration_tests/snapshotenvironmentbinding_controller_test.go
@@ -20,7 +20,6 @@ import (
 	"github.com/redhat-appstudio/application-api/api/v1alpha1"
 	rapi "github.com/redhat-appstudio/remote-secret/api/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
 )
 
 var _ = Describe("SnapshotEnvironmentBinding", func() {
@@ -127,62 +126,5 @@ var _ = Describe("SnapshotEnvironmentBinding", func() {
 			})
 		})
 
-	})
-
-	Describe("Removes the target for RemoteSecrets with environment removal ", func() {
-
-		env := StandardRemoteEnvironment("test-remote-env")
-
-		testSetup := TestSetup{
-			ToCreate: TestObjects{
-				RemoteSecrets: []*rapi.RemoteSecret{{
-					ObjectMeta: metav1.ObjectMeta{
-						GenerateName: "remove-target-remotesecret-",
-						Namespace:    env.Namespace,
-						Labels:       map[string]string{"appstudio.redhat.com/application": "test-app", "appstudio.redhat.com/environment": env.Name},
-					},
-					Spec: rapi.RemoteSecretSpec{
-						Secret: rapi.LinkableSecretSpec{
-							Name: "test-remote-secret",
-							Type: "Opaque",
-						},
-					},
-				}},
-				Environments: []*v1alpha1.Environment{env},
-				SnapshotEnvBindings: []*v1alpha1.SnapshotEnvironmentBinding{{
-					ObjectMeta: metav1.ObjectMeta{
-						GenerateName: "create-target-snapshotbinding-",
-						Namespace:    env.Namespace,
-						Labels:       map[string]string{"appstudio.redhat.com/application": "test-app", "appstudio.redhat.com/environment": env.Name},
-					},
-					Spec: v1alpha1.SnapshotEnvironmentBindingSpec{
-						Application: "test-app",
-						Environment: env.Name,
-						Components:  []v1alpha1.BindingComponent{},
-					},
-				}},
-			},
-			Behavior: ITestBehavior{},
-		}
-		BeforeEach(func() {
-			testSetup.BeforeEach(nil)
-		})
-
-		var _ = AfterEach(func() {
-			testSetup.AfterEach()
-		})
-
-		It("have the target deleted", func() {
-			//check target is there
-			Expect(testSetup.InCluster.RemoteSecrets[0].Spec.Targets).ToNot(BeEmpty())
-			// delete environment
-			Expect(ITest.Client.Delete(ITest.Context, env)).To(Succeed())
-			// check target is gone
-			Eventually(func(g Gomega) {
-				rs := &rapi.RemoteSecret{}
-				g.Expect(ITest.Client.Get(ITest.Context, types.NamespacedName{Name: testSetup.InCluster.RemoteSecrets[0].Name, Namespace: testSetup.InCluster.RemoteSecrets[0].Namespace}, rs)).To(Succeed())
-				g.Expect(rs.Spec.Targets).To(BeEmpty())
-			}).Should(Succeed())
-		})
 	})
 })
